chore(gatsby): Convert develop-static to TypeScript
